Data/SQLite: Exception messages contain duplicate text #2012

+++ Data/SQLite/src/Extractor.cpp
@@ -228,7 +228,7 @@ bool Extractor::isNull(std::size_t pos, std::size_t)
 	if (!_nulls[pos].first)
 	{
 		_nulls[pos].first = true;
-		_nulls[pos].second = (SQLITE_NULL == sqlite3_column_type(_pStmt, pos));
+		_nulls[pos].second = (SQLITE_NULL == sqlite3_column_type(_pStmt, static_cast<int>(pos)));
 	}
 	
 	return _nulls[pos].second;
